case_study/8_financial_document_analysis/requirements.txt
llama-index
llama-index-core
llama-index-embeddings-huggingface
llama-index-llms-gemini
python-dotenv
redis